% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie2miscdata.R
\docType{package}
\name{ie2miscdata}
\alias{ie2miscdata}
\alias{ie2miscdata-package}
\title{ie2miscdata: Irucka Embry's Miscellaneous USGS Data Collection}
\description{
ie2miscdata provides various data sets [USGS Parameter codes with fixed
  values, USGS global time zone codes, and US Air Force Global Engineering
  Weather Data] created by Irucka Embry while he was a Cherokee Nation
  Technology Solutions (CNTS) United States Geological Survey (USGS)
  Contractor and/or USGS employee.
}

