% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_00041.R
\docType{data}
\name{pmcode_00041}
\alias{pmcode_00041}
\title{00041 Weather}
\format{A data frame with 54 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_00041
}
\description{
A table containing the USGS Weather parameter codes.
}
\examples{
pmcode_00041
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

