% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_00067.R
\docType{data}
\name{pmcode_00067}
\alias{pmcode_00067}
\title{00067 Tide stage, code}
\format{
A data frame with 315 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_00067
}
\description{
A table containing the USGS Tide stage, code parameter codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
