% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_01315.R
\docType{data}
\name{pmcode_01315}
\alias{pmcode_01315}
\title{01315 Sludge: Floating (Severity)}
\format{A data frame with 5 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_01315
}
\description{
A table containing the USGS Sludge: Floating (Severity) parameter codes.
}
\examples{
pmcode_01315
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

