% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_01335.R
\docType{data}
\name{pmcode_01335}
\alias{pmcode_01335}
\title{01335 Sewage Solids, Fresh, Floating (Severity)}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_01335
}
\description{
A table containing the USGS Sewage Solids, Fresh, Floating (Severity)
parameter codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
