% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_01355.R
\docType{data}
\name{pmcode_01355}
\alias{pmcode_01355}
\title{01355 Ice Cover, Floating Or Solid (Severity)}
\format{A data frame with 5 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_01355
}
\description{
A table containing the USGS Ice Cover, Floating Or Solid (Severity)
parameter codes.
}
\examples{
pmcode_01355
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

