% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_04117.R
\docType{data}
\name{pmcode_04117}
\alias{pmcode_04117}
\title{04117 Tether Line Used For Collecting Sample (Yes=1) Codes}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_04117
}
\description{
A table containing the USGS Tether Line Used For Collecting Sample (Yes=1)
Codes parameter codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
