% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_50276.R
\docType{data}
\name{pmcode_50276}
\alias{pmcode_50276}
\title{50276 Filter Type, Code}
\format{A data frame with 14 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_50276
}
\description{
A table containing the USGS Filter Type, Code parameter codes.
}
\examples{
pmcode_50276
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

