% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_50280.R
\docType{data}
\name{pmcode_50280}
\alias{pmcode_50280}
\title{50280 Water Samples, Code}
\format{A data frame with 28 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_50280
}
\description{
A table containing the USGS Water Samples, Code parameter codes.
}
\examples{
pmcode_50280
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

