% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_62955.R
\docType{data}
\name{pmcode_62955}
\alias{pmcode_62955}
\title{62955 Sample Matrix, Code}
\format{
A data frame with 3 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_62955
}
\description{
A table containing the USGS Sample Matrix, Code parameter codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
