% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_71996.R
\docType{data}
\name{pmcode_71996}
\alias{pmcode_71996}
\title{71996 Water Use, Secondary (Codes)}
\format{A data frame with 131 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_71996
}
\description{
A table containing the USGS Water Use, Secondary (Codes) parameter codes.
}
\examples{
pmcode_71996
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

