% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_72006.R
\docType{data}
\name{pmcode_72006}
\alias{pmcode_72006}
\title{72006 Sampling Condition (Codes)}
\format{A data frame with 46 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_72006
}
\description{
A table containing the USGS Sampling Condition (Codes) parameter codes.
}
\examples{
pmcode_72006
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

