% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_74200.R
\docType{data}
\name{pmcode_74200}
\alias{pmcode_74200}
\title{74200 Sample Preservation Method, (Codes)}
\format{
A data frame with 119 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_74200
}
\description{
A table containing the USGS Sample Preservation Method, (Codes) parameter
codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
