% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_82305.R
\docType{data}
\name{pmcode_82305}
\alias{pmcode_82305}
\title{82305 Atmospheric Deposition Type Bulk, (Codes)}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_82305
}
\description{
A table containing the USGS Atmospheric Deposition Type Bulk, (Codes)
parameter codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
