% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_82398.R
\docType{data}
\name{pmcode_82398}
\alias{pmcode_82398}
\title{82398 Sampling Method (Codes)}
\format{A data frame with 63 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_82398
}
\description{
A table containing the USGS Sampling Method (Codes) parameter codes.
}
\examples{
pmcode_82398
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

