% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_84143.R
\docType{data}
\name{pmcode_84143}
\alias{pmcode_84143}
\title{84143 Well Purging Condition (Codes)}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_84143
}
\description{
A table containing the USGS Well Purging Condition (Codes) parameter codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
