% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_84145.R
\docType{data}
\name{pmcode_84145}
\alias{pmcode_84145}
\title{84145 Project Component (Codes)}
\format{A data frame with 7 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_84145
}
\description{
A table containing the USGS Project Component (Codes) parameter codes.
}
\examples{
pmcode_84145
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

