% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_84146.R
\docType{data}
\name{pmcode_84146}
\alias{pmcode_84146}
\title{84146 Land Use, Predominant, Within 200 Feet Of Well, (Codes)}
\format{A data frame with 15 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_84146
}
\description{
A table containing the USGS Land Use, Predominant, Within 200 Feet Of Well,
(Codes) parameter codes.
}
\examples{
pmcode_84146
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

