% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_84164.R
\docType{data}
\name{pmcode_84164}
\alias{pmcode_84164}
\title{84164 Sampler Type, (Codes)}
\format{A data frame with 121 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_84164
}
\description{
A table containing the USGS Sampler Type, (Codes) parameter codes.
}
\examples{
pmcode_84164
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

