% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_91112.R
\docType{data}
\name{pmcode_91112}
\alias{pmcode_91112}
\title{91112 Latitude/Longitude Horizontal Datum}
\format{A data frame with 7 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_91112
}
\description{
A table containing the USGS Latitude/Longitude Horizontal Datum parameter
codes.
}
\examples{
pmcode_91112
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

