% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_91113.R
\docType{data}
\name{pmcode_91113}
\alias{pmcode_91113}
\title{91113 Latitude/Longitude Measurement Method}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_91113
}
\description{
A table containing the USGS Latitude/Longitude Measurement Method parameter
codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
