% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_91114.R
\docType{data}
\name{pmcode_91114}
\alias{pmcode_91114}
\title{91114 Latitude/Longitude Coordinate Accuracy}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_91114
}
\description{
A table containing the USGS Latitude/Longitude Coordinate Accuracy parameter
codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
