% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99101.R
\docType{data}
\name{pmcode_99101}
\alias{pmcode_99101}
\title{99101 Blank, Source Of Solution, Fixed Value Code}
\format{A data frame with 25 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_99101
}
\description{
A table containing the USGS Blank, Source Of Solution, Fixed Value Code
parameter codes.
}
\examples{
pmcode_99101
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

