% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99102.R
\docType{data}
\name{pmcode_99102}
\alias{pmcode_99102}
\title{99102 Blank, Type Of Sample, Fixed Value Code}
\format{A data frame with 13 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_99102
}
\description{
A table containing the USGS Blank, Type Of Sample, Fixed Value Code
parameter codes.
}
\examples{
pmcode_99102
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

