% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99106.R
\docType{data}
\name{pmcode_99106}
\alias{pmcode_99106}
\title{99106 Spike, Type, Fixed Value Code}
\format{A data frame with 5 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_99106
}
\description{
A table containing the USGS Spike, Type, Fixed Value Code parameter codes.
}
\examples{
pmcode_99106
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

