% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99111.R
\docType{data}
\name{pmcode_99111}
\alias{pmcode_99111}
\title{99111 Quality Assurance Data Type Associated With Sample, Code}
\format{A data frame with 9 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_99111
}
\description{
A table containing the USGS Quality Assurance Data Type Associated With
Sample, Code parameter codes.
}
\examples{
pmcode_99111
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

