% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99112.R
\docType{data}
\name{pmcode_99112}
\alias{pmcode_99112}
\title{99112 Sulfide, Water, Filtered, Field, milligrams Per Liter}
\format{A data frame with 10 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_99112
}
\description{
A table containing the USGS Sulfide, Water, Filtered, Field, milligrams Per
Liter parameter codes.
}
\examples{
pmcode_99112
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

