% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99329.R
\docType{data}
\name{pmcode_99329}
\alias{pmcode_99329}
\title{99329 Coliphage, somatic, E. coli C-host, 2-stepenrichment presence/absence per 1 liter}
\format{A data frame with 2 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_99329
}
\description{
A table containing the USGS Coliphage, somatic, E. coli C-host,
2-stepenrichment presence/absence per 1 liter parameter codes.
}
\examples{
pmcode_99329
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

