% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99335.R
\docType{data}
\name{pmcode_99335}
\alias{pmcode_99335}
\title{99335 ColipgeF-spec, famp, 2-step,pres(1) abs(2)/ 1 L}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}
}
\usage{
pmcode_99335
}
\description{
A table containing the USGS ColipgeF-spec, famp, 2-step,pres(1) abs(2)/ 1 L
parameter codes.
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{https://help.waterdata.usgs.gov/codes-and-parameters/}.
}
\keyword{datasets}
