% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_71999.R
\docType{data}
\name{pmcode_71999}
\alias{pmcode_71999}
\title{71999 Sample Purpose (Codes)}
\format{A data frame with 28 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_71999
}
\description{
A table containing the USGS Sample Purpose (Codes) parameter codes.
}
\examples{
pmcode_71999
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

