% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_72005.R
\docType{data}
\name{pmcode_72005}
\alias{pmcode_72005}
\title{72005 Sample Source (Codes)}
\format{A data frame with 101 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_72005
}
\description{
A table containing the USGS Sample Source (Codes) parameter codes.
}
\examples{
pmcode_72005
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

