% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_82923.R
\docType{data}
\name{pmcode_82923}
\alias{pmcode_82923}
\title{82923 Atmospheric Deposition Type Wet, (Codes)}
\format{A data frame with 7 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_82923
}
\description{
A table containing the USGS Atmospheric Deposition Type Wet, (Codes)
parameter codes.
}
\examples{
pmcode_82923
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

