% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_84060.R
\docType{data}
\name{pmcode_84060}
\alias{pmcode_84060}
\title{84060 Topography, Physiographic Setting (Codes)}
\format{A data frame with 18 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_84060
}
\description{
A table containing the USGS Topography, Physiographic Setting (Codes)
parameter codes.
}
\examples{
pmcode_84060
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

