% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcode_99595.R
\docType{data}
\name{pmcode_99595}
\alias{pmcode_99595}
\title{99595 Total coliform, ColilrtPA, wtrprs/abs /1L}
\format{A data frame with 2 rows and 3 variables:
\describe{
\item{Parameter Code}{USGS Parameter Code}
\item{Fixed Value}{Fixed Value}
\item{Fixed Text}{Fixed Text}
}}
\usage{
pmcode_99595
}
\description{
A table containing the USGS Total coliform, ColilrtPA, wtrprs/abs /1L
parameter codes.
}
\examples{
pmcode_99595
}
\references{
This data is from Table 26. Parameter codes with fixed values (USGS Water Quality Samples for USA: Sample Data). See \url{http://waterdata.usgs.gov/usa/nwis/qwdata?codes_table26_help}.
}

