% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_codes.R
\docType{data}
\name{tz_codes}
\alias{tz_codes}
\title{Timezone (tz) codes}
\format{A data frame with 54 rows and 7 variables:
\describe{
\item{Tz Cd}{USGS Parameter Code}
\item{Tz Nm}{Fixed Value}
\item{Tz Ds}{Fixed Text}
\item{Tz Utc Offset Tm}{USGS Parameter Code}
\item{Tz Dst Cd}{Fixed Value}
\item{Tz Dst Nm}{Fixed Text}
\item{Tz Dst Utc Offset Tm}{Fixed Text}
}}
\usage{
tz_codes
}
\description{
A table containing the worldwide timezones and GMT offsets.
codes.
}
\examples{
tz_codes
}
\references{
This data is from National Water Information System: Help System Time Zone Codes. See \url{http://help.waterdata.usgs.gov/code/tz_query?fmt=html}.
}

