% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-weather_results.R
\docType{data}
\name{weather_results}
\alias{weather_results}
\title{Manning's n for Channels}
\format{A data frame with 805 rows and 7 variables:
\describe{
\item{Type of Channel and Description}{Type of channel name and any
descriptive information}
\item{Minimum n}{Minimum n value}
\item{Normal n}{Normal n value}
\item{Maximum n}{Maximum n value}
\item{Minimum n}{Minimum n value}
\item{Normal n}{Normal n value}
\item{Maximum n}{Maximum n value}
}}
\usage{
weather_results
}
\description{
A table containing the type of channel and description along with the
minimum, normal, and maximum value of n, if it exists. n is the "Gauckler-
Manning coefficient (commonly called Manning's n) and it's units are
(T/[L1/3]; s/[ft1/3]; s/[m1/3])". Source: Manning formula: From Wikipedia,
the free encyclopedia. See \url{https://en.wikipedia.org/wiki/Manning_formula}.
}
\examples{
weather_results
}
\references{
This data is from FishXing Version 3.0 Beta (2006) by Michael Furniss,
 Michael Love, Susan Firor, Kathleen Moynan, Antonio Llanos, Jeff Guntle,
 and Robert Gubernick.
 See \url{http://www.fsl.orst.edu/geowater/FX3/help/8_Hydraulic_Reference/Mannings_n_Tables.htm}.
 The original data source is Ven Te Chow, \emph{Open-Channel Hydraulics},
 New York City, York: McGraw-Hill, 1959.
}

